+++ package.json
@@ -65,10 +65,12 @@
     "depcruise": "node ./bin/dependency-cruise.js --validate -- src bin test configs types",
     "depcruise:explain": "node ./bin/dependency-cruise.js --output-type err-long --validate -- src bin test configs types",
     "depcruise:graph:dir": "node ./bin/dependency-cruise.js --validate --include-only '(^src|^[^/]+(js|json)$)' --output-type ddot bin src | dot -T png > tmp_deps.png",
-    "depcruise:graph:doc": "npm-run-all depcruise:graph:doc:dot depcruise:graph:doc:html depcruise:graph:doc:png",
+    "depcruise:graph:doc": "npm-run-all depcruise:graph:doc:dot depcruise:graph:doc:html depcruise:graph:doc:png depcruise:graph:doc:schema depcruise:graph:doc:types",
     "depcruise:graph:doc:dot": "node ./bin/dependency-cruise.js --validate --include-only '^(src|bin)' --output-type dot bin src --output-to tmp_graph_deps.dot",
     "depcruise:graph:doc:html": "cat tmp_graph_deps.dot | dot -T svg -Gsplines=ortho | cat docs/assets/depcruise-graph-head.snippet.html - docs/assets/depcruise-graph-foot.snippet.html > docs/dependency-cruiser-dependency-graph.html",
     "depcruise:graph:doc:png": "cat tmp_graph_deps.dot | dot -Gdpi=192 -Gsplines=ortho -T png | pngquant - > doc/real-world-samples/dependency-cruiser-without-node_modules.png",
+    "depcruise:graph:doc:schema": "cd utl/schema && node ../../bin/dependency-cruise.js --validate ../../.dependency-cruiser.json --output-type dot . | dot -Grankdir=LR -Gsplines=ortho -T svg > ../overview.svg && cd -",
+    "depcruise:graph:doc:types": "cd types && node ../bin/dependency-cruise.js --ts-pre-compilation-deps --validate ../.dependency-cruiser.json --output-type dot . | dot -Grankdir=TD -Gsplines=ortho -T svg > overview.svg && cd -",
     "depcruise:graph:osage": "node ./bin/dependency-cruise.js --validate --include-only '(^src|^[^/]+(js|json)$)' --output-type dot bin src | osage -Gpack=16 -GpackMode=array2 -T svg > tmp_deps.svg",
     "depcruise:graph:fdp": "node ./bin/dependency-cruise.js --validate --include-only '^src' --output-type dot bin src | fdp -GK=0.1 -Gsplines=ortho -T svg > tmp_deps.svg",
     "depcruise:report": "node ./bin/dependency-cruise.js --output-type err-html --validate -- src bin test configs types > dependency-violations.html",
