+++ src/schema/cruise-result.schema.json
@@ -14,18 +14,6 @@
     }
   },
   "definitions": {
-    "ModuleSystemType": {
-      "type": "string",
-      "enum": ["cjs", "amd", "es6", "tsd"]
-    },
-    "ModuleSystemsType": {
-      "type": "array",
-      "description": "List of module systems to cruise. Defaults to [amd, cjs, es6]",
-      "items": {
-        "type": "string",
-        "enum": ["amd", "cjs", "es6", "tsd"]
-      }
-    },
     "SeverityType": {
       "type": "string",
       "description": "How severe a violation of a rule is. The 'error' severity will make some reporters return a non-zero exit code, so if you want e.g. a build to stop when there's a rule violated: use that.",
@@ -283,6 +271,32 @@
         }
       }
     },
+    "ModuleSystemType": {
+      "type": "string",
+      "enum": ["cjs", "es6", "amd", "tsd"]
+    },
+    "ModuleSystemsType": {
+      "type": "array",
+      "description": "List of module systems to cruise. Defaults to [amd, cjs, es6]",
+      "items": {
+        "$ref": "#/definitions/ModuleSystemType"
+      }
+    },
+    "OutputType": {
+      "type": "string",
+      "enum": [
+        "json",
+        "html",
+        "dot",
+        "ddot",
+        "csv",
+        "err",
+        "err-long",
+        "err-html",
+        "teamcity",
+        "anon"
+      ]
+    },
     "OptionsUsedType": {
       "type": "object",
       "description": "the (command line) options used to generate the dependency-tree",
@@ -494,21 +508,6 @@
           "description": "The circular path if the violation was about circularity"
         }
       }
-    },
-    "OutputType": {
-      "type": "string",
-      "enum": [
-        "json",
-        "html",
-        "dot",
-        "ddot",
-        "csv",
-        "err",
-        "err-long",
-        "err-html",
-        "teamcity",
-        "anon"
-      ]
     },
     "RuleSummaryType": {
       "type": "object",
