+++ utl/README.md
@@ -1,6 +1,7 @@
-The [schema](schema) folder contains the sources for both json schemas. They're split up 
-for ease of maintenance. To generate the schemas after updating any of the stuff in there
-run:
+The [schema](schema) folder contains the sources for both json schemas ([configuration](../src/schemas/configuration.schema.json) and [cruise-result](../src/schemas/cruise-result.schema.json) . They're split up 
+for ease of maintenance). 
+
+To generate the schemas from their sources run:
 
 ```sh
 node utl/generate-schemas.utl.js
@@ -9,5 +10,3 @@ node utl/generate-schemas.utl.js
 This is done on each build as well
 
 ![overview](overview.svg)
-
-<img width="712" alt="original" src="overview.png">
