+++ utl/schema/dependency-type.js
@@ -1,0 +1,23 @@
+module.exports = {
+  definitions: {
+    DependencyType: {
+      type: "string",
+      enum: [
+        "aliased",
+        "core",
+        "deprecated",
+        "local",
+        "localmodule",
+        "npm",
+        "npm-bundled",
+        "npm-dev",
+        "npm-no-pkg",
+        "npm-optional",
+        "npm-peer",
+        "npm-unknown",
+        "undetermined",
+        "unknown"
+      ]
+    }
+  }
+};
