+++ utl/schema/module-system-type.js
@@ -1,0 +1,8 @@
+module.exports = {
+  definitions: {
+    ModuleSystemType: {
+      type: "string",
+      enum: ["cjs", "es6", "amd", "tsd"]
+    }
+  }
+};
