+++ utl/schema/module-systems-type.js
@@ -1,0 +1,13 @@
+const moduleSystemType = require("./module-system-type");
+
+module.exports = {
+  definitions: {
+    ...moduleSystemType.definitions,
+    ModuleSystemsType: {
+      type: "array",
+      description:
+        "List of module systems to cruise. Defaults to [amd, cjs, es6]",
+      items: { $ref: "#/definitions/ModuleSystemType" }
+    }
+  }
+};
