+++ utl/schema/output-type.js
@@ -1,0 +1,19 @@
+module.exports = {
+  definitions: {
+    OutputType: {
+      type: "string",
+      enum: [
+        "json",
+        "html",
+        "dot",
+        "ddot",
+        "csv",
+        "err",
+        "err-long",
+        "err-html",
+        "teamcity",
+        "anon"
+      ]
+    }
+  }
+};
